move few utils functions to textUtils file
