+++ jsapp/js/account/accountSettingsRoute.tsx
@@ -6,7 +6,8 @@ import sessionStore from 'js/stores/session';
 import './accountSettings.scss';
 import Checkbox from '../components/common/checkbox';
 import TextBox from '../components/common/textBox';
-import {addRequiredToLabel, notify, stringToColor} from '../utils';
+import {notify, stringToColor} from 'js/utils';
+import {addRequiredToLabel} from 'js/textUtils';
 import envStore from '../envStore';
 import WrappedSelect from '../components/common/wrappedSelect';
 import {dataInterface} from '../dataInterface';
