+++ jsapp/js/actions.es6
@@ -17,10 +17,8 @@ import submissionsActions from './actions/submissions';
 import formMediaActions from './actions/mediaActions';
 import exportsActions from './actions/exportsActions';
 import dataShareActions from './actions/dataShareActions';
-import {
-  notify,
-  replaceSupportEmail,
-} from 'utils';
+import {notify} from 'js/utils';
+import {replaceSupportEmail} from 'js/textUtils';
 
 // Configure Reflux
 Reflux.use(RefluxPromise(window.Promise));
