+++ jsapp/js/components/modalForms/languageForm.es6
@@ -2,10 +2,8 @@ import React from 'react';
 import autoBind from 'react-autobind';
 import bem from 'js/bem';
 import TextBox from 'js/components/common/textBox';
-import {
-  getLangAsObject,
-  toTitleCase
-} from 'utils';
+import {getLangAsObject} from 'js/utils';
+import {toTitleCase} from 'js/textUtils';
 
 /*
 Properties:
