+++ jsapp/js/utils.ts
@@ -14,7 +14,6 @@ import {Cookies} from 'react-cookie';
 // importing whole constants, as we override ROOT_URL in tests
 import constants from 'js/constants';
 import type Raven from 'raven';
-import envStore from 'js/envStore';
 
 export const LANGUAGE_COOKIE_NAME = 'django_language';
 
@@ -210,33 +209,6 @@ export const log = (function () {
 })();
 window.log = log;
 
-const ORIGINAL_SUPPORT_EMAIL = 'help@kobotoolbox.org';
-
-/**
- * Replaces the hardcoded email string (coming from transifex translation) with
- * the one from the `/environment` endpoint.
- */
-export function replaceSupportEmail(str: string): string {
-  if (
-    typeof envStore.data.support_email === 'string' &&
-    envStore.data.support_email.length !== 0
-  ) {
-    return str.replace(ORIGINAL_SUPPORT_EMAIL, envStore.data.support_email);
-  } else {
-    return str;
-  }
-}
-
-// returns an HTML string where [bracket] notation is replaced with a hyperlink
-export function replaceBracketsWithLink(str: string, url?: string): string {
-  const bracketRegex = /\[([^\]]+)\]/g;
-  if (!url) {
-    return str.replace(bracketRegex, '$1');
-  }
-  const linkHtml = `<a href="${url}" target="_blank">$1</a>`;
-  return str.replace(bracketRegex, linkHtml);
-}
-
 export function currentLang(): string {
   return cookies.get(LANGUAGE_COOKIE_NAME) || 'en';
 }
@@ -276,14 +248,6 @@ export function getLangString(obj: LangObject): string | undefined {
   } else {
     return undefined;
   }
-}
-
-export function addRequiredToLabel(label: string, isRequired = true): string {
-  if (!isRequired) {
-    return label;
-  }
-  const requiredTemplate = t('##field_label## (required)');
-  return requiredTemplate.replace('##field_label##', label);
 }
 
 export function stringToColor(str: string, prc?: number) {
@@ -367,12 +331,6 @@ export function renderCheckbox(id: string, label: string, isImportant = false) {
   return `<div class="alertify-toggle checkbox ${additionalClass}"><label class="checkbox__wrapper"><input type="checkbox" class="checkbox__input" id="${id}" data-cy="checkbox"><span class="checkbox__label">${label}</span></label></div>`;
 }
 
-export function hasLongWords(text: string, limit = 25): boolean {
-  const textArr = text.split(' ');
-  const maxLength = Math.max(...textArr.map((el) => el.length));
-  return maxLength >= limit;
-}
-
 export function hasVerticalScrollbar(element: HTMLElement): boolean {
   return element.scrollHeight > element.offsetHeight;
 }
@@ -404,10 +362,6 @@ export function getScrollbarWidth(): number {
   }
 
   return scrollbarWidth;
-}
-
-export function toTitleCase(str: string): string {
-  return str.replace(/(^|\s)\S/g, (t) => t.toUpperCase());
 }
 
 export function launchPrinting() {
